requirements.txt
# --- Web Framework & Server ---
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
python-multipart>=0.0.6

# --- Database & ORM ---
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0

# --- Data Processing & Excel ---
pandas>=2.0.0
openpyxl>=3.1.0

# --- Utilities ---
python-slugify>=8.0.0
pydantic>=2.0.0
Pillow>=10.0.0